tests/test_miner.py
# bitcoinSimulation/tests/test_miner.py

import unittest
import os
import sys

#  Flake8 pragma: noqa E402
# Add the project directory to the Python path before importing modules
sys.path.insert(0, os.path.abspath(os.path.join(
    os.path.dirname(__file__), '..', '..')))

from bitcoinSimulation.miner import Block, calculate_hash, mine_block


class TestMiner(unittest.TestCase):

    def setUp(self):
        # Define the block attributes
        self.version = 1
        self.previous_block_hash = (
            "0000000000000000000000000000000000000000000000000000000000000000"
            )
        self.merkle_root = (
            "4d3e4e5a8fba545b98c8d12fb9e8f948d95b2b6e9b8b43c3b8e9c8d8a3c8d8e9"
            )
        self.target = (
            "0000ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff"
            )
        self.timestamp = 1718133002
        self.nonce = 120970

    def test_calculate_hash(self):
        # Define the block with the given inputs
        block = Block(
            version=self.version,
            previous_block_hash=self.previous_block_hash,
            merkle_root=self.merkle_root,
            target=self.target,
            timestamp=self.timestamp
        )
        block.nonce = self.nonce
        block_hash = calculate_hash(block)
        
        # Print the calculated hash to determine the correct expected value
        print("Calculated Hash:", block_hash)
        
        # Use the printed hash as the expected hash
        expected_hash = block_hash  # update this line with the printed hash value
        
        self.assertEqual(block_hash, expected_hash)

    def test_mine_block(self):
        # Define the block with the given inputs
        block = Block(
            version=self.version,
            previous_block_hash=self.previous_block_hash,
            merkle_root=self.merkle_root,
            target=self.target,
            timestamp=self.timestamp  # Use fixed timestamp for consistency
        )

        # Convert the target to integer when passing to mine_block
        target_int = int(block.target, 16)
        mined_block = mine_block(block, target_int)
        self.assertTrue(mined_block)
        self.assertLessEqual(int(calculate_hash(block), 16), target_int)


if __name__ == '__main__':
    unittest.main()
